requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
ollama==0.6.1
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
langchain-ollama==0.3.10

-e .